+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Convert svgs from px to mm
 - Add 12 dB/oct filter slope option to Torus (default is 6dB/oct)
 - Add LFO option to Planck functions in Energy
+- Make Fate polyphonic
 
 
 ### 1.1.1 (2020-03-24)